updated seeds and README
