--- .env
@@ -1,3 +1,0 @@
-DB_NAME=stagename_db
-DB_PASSWORD=Abc123
-DB_USER=root
