+++ README.md
@@ -1,1 +1,111 @@
 # Project-2
+
+
+## Team
+
+Tobias Virgona(engineer, back end developer ,de-bugger) [Github](https://github.com/virgona)
+Rosemary Forster(engineer, front end developer) [Github](https://github.com/rosemaryjf)
+Alex Mastroianni(engineer, back end developer) [Github](https://github.com/AlexMastroianni/)
+Chris Gatt(engineer, back end developer) [Github](https://github.com/ChristopherGatt/)
+
+
+## About
+
+Our project is called MusicFwends! A simple social media website that allows users to interact and communicate with a focal point on music. Users can post their music tastes and views publicly whilst other users can view, comment and react accordingly.
+
+
+## Purpose
+
+The purpose of this project is to collaborate with a team of developers, coming together whilst using multiple technologies in conjuction to create a real-world full-stack application that you’ll be able to showcase to potential employers.
+
+
+## Requirements
+
+Use Node.js and Express.js to create a RESTful API.
+
+Use Handlebars.js as the templating engine.
+
+Use MySQL and the Sequelize ORM for the database.
+
+Have both GET and POST routes for retrieving and adding new data.
+
+Be deployed using Heroku (with data).
+
+Use at least one new library, package, or technology that we haven’t discussed.
+
+Have a polished UI.
+
+Be responsive.
+
+Be interactive (i.e., accept and respond to user input).
+
+Have a folder structure that meets the MVC paradigm.
+
+Include authentication (express-session and cookies).
+
+Protect API keys and sensitive information with environment variables.
+
+Have a clean repository that meets quality coding standards (file structure, naming conventions, follows best practices for class/id naming conventions, indentation, quality comments, etc.).
+
+Have a quality README (with unique name, description, technologies used, screenshot, and link to deployed application).
+
+Finally, you must add your project to the portfolio that you created in Module 2.
+
+
+## Technologies used
+
+Node.js
+
+Express.js
+
+HandleBars
+
+MySQL
+
+Sequelize
+
+Heroku
+
+eslint
+
+Visual Code Studio
+
+Bootstrap
+
+
+## Functionality
+
+1. The user is first taken to a login page where they will enter an email and password to either login or sign up. 
+
+2. The user will determine this by clicking on the corresponding button (login, sign up) located at the bottom of tha page
+
+3. The password strength is displayed (weak, medium, strong)
+
+4. Once the user has logged in, they will be re-directed to the main page
+
+5. Here the user can post content for public viewing using the "create post" button
+
+6. Users can interact with post either by commenting or reacting
+
+7. Users can can delete their post using the "delete post" button
+
+8. If the page is reloaded the data will be saved
+
+
+## User Interface Screenshots
+
+![MusicFwends login/signup page]()
+
+![MusicFwends main page]()
+
+![MusicFwends in use]()
+
+
+## Github
+
+The repository can be found [here](https://github.com/AlexMastroianni/Project-2)
+
+
+## Deployed Application
+
+The Deployed Appliction can be found [here!]()
