+++ models/User.js
@@ -30,19 +30,13 @@ User.init(
       validate: {
         len: [10],
       },
-
-      username: {
-        type: DataTypes.STRING,
-        allowNull: false,
-      },
-      profile_avatar: {
-        type: DataTypes.STRING,
-      },
-      posts: {
-        type: DataTypes.STRING('long'),
-        allowNull: true,
-      },
-
+    },
+    username: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    profile_avatar: {
+      type: DataTypes.STRING,
     },
   },
   {
