--- seeds/User-seeds.js
@@ -1,43 +1,0 @@
-const { User } = require('../models');
-
-const userData = [
-  {
-    id: 1,
-    email: 'feelincute@email.com',
-    password: 'secret123',
-    username: 'poop',
-    profile_avatar: 'https://via.placeholder.com/150x150',
-  },
-  {
-    id: 2,
-    email: 'feelin cutey',
-    password: 'secret123',
-    username: 'poop',
-    profile_avatar: 'https://via.placeholder.com/150x150',
-  },
-  {
-    id: 3,
-    email: 'feelin cuteyy',
-    password: 'secret123s',
-    username: 'poops',
-    profile_avatar: 'https://via.placeholder.com/150x150',
-  },
-  {
-    id: 4,
-    email: 'feelin cutessy',
-    password: 'secret123ss',
-    username: 'poopss',
-    profile_avatar: 'https://via.placeholder.com/150x150',
-  },
-  {
-    id: 5,
-    email: 'feelin cut',
-    password: 'secret12',
-    username: 'poo',
-    profile_avatar: 'https://via.placeholder.com/150x150',
-  },
-];
-
-// const seedProducts = () => PostTable.bulkCreate(postData);
-
-module.exports = userData;
