+++ seeds/postSeeds.js
@@ -1,36 +1,34 @@
-// const { PostTable } = require('../models');
-
 const postData = [
   {
     id: 1,
     content: 'feelin cute',
-    author_id: 1,
+    author_id: 1
     //reaction: "might delete later",
   },
   {
     id: 2,
     content: 'Harry <3',
     author_id: 2,
-    reaction: 'so cool!',
+    reaction: 'so cool!'
   },
   {
     id: 3,
     content: 'Linkin Park GOATED',
-    author_id: 3,
+    author_id: 3
     //reaction: "OMG",
   },
   {
     id: 4,
     content: 'Brittany is bad',
-    author_id: 4,
+    author_id: 4
     //reaction: "hard",
   },
   {
     id: 5,
     content: 'yeeew',
-    author_id: 5,
+    author_id: 5
     //reaction: "yooo",
-  },
+  }
 ];
 
 //const seedPostData = () => PostTable.bulkCreate(postData);
