+++ seeds/seed.js
@@ -1,16 +1,16 @@
 const sequelize = require('../config/connection');
 const { User, Comment, Post } = require('../models');
 
-const userData = require('./User-seeds.js');
-const postData = require('./PostTable-seeds.js');
-const commentData = require('./CommentTable-seeds.js');
+const userData = require('./userSeeds.js');
+const postData = require('./postSeeds.js');
+const commentData = require('./commentSeeds.js');
 
 const seedDatabase = async () => {
   await sequelize.sync({ force: true });
 
-  const users = await User.bulkCreate(userData, {
+  await User.bulkCreate(userData, {
     individualHooks: true,
-    returning: true,
+    returning: true
   });
 
   for (const post of postData) {
