+++ seeds/userSeeds.js
@@ -1,0 +1,41 @@
+const userData = [
+  {
+    id: 1,
+    email: 'feelincute@gmail.com',
+    password: 'secret123',
+    username: 'poop',
+    profile_avatar: 'https://via.placeholder.com/150x150'
+  },
+  {
+    id: 2,
+    email: 'feelincutey@gmail.com',
+    password: 'secret123',
+    username: 'poopi',
+    profile_avatar: 'https://via.placeholder.com/150x150'
+  },
+  {
+    id: 3,
+    email: 'feelincuteyy@gmail.com',
+    password: 'secret123s',
+    username: 'poops',
+    profile_avatar: 'https://via.placeholder.com/150x150'
+  },
+  {
+    id: 4,
+    email: 'feelincutessy@gmail.com',
+    password: 'secret123ss',
+    username: 'poopss',
+    profile_avatar: 'https://via.placeholder.com/150x150'
+  },
+  {
+    id: 5,
+    email: 'feelincut@gmail.com',
+    password: 'secret12',
+    username: 'wee',
+    profile_avatar: 'https://via.placeholder.com/150x150'
+  }
+];
+
+// const seedProducts = () => PostTable.bulkCreate(postData);
+
+module.exports = userData;
